Fixed login and style issues (#9)
* Fixed UI style and login success redirection related issues
