+++ ideaforce/src/main/java/com/emc/ideaforce/utils/Utils.java
@@ -5,7 +5,7 @@ public final class Utils {
     private Utils() {
     }
 
-    public static final String PWD_PRIVELEGE = "CHANGE_PASSWORD_PRIVILEGE";
+    public static final String CP_PRIVILEGE = "CP_PRIVILEGE";
     public static final String REG_USER_ROLE = "ROLE_USER";
     public static final String ADMIN_ROLE = "ROLE_ADMIN";
