+++ ideaforce/src/main/resources/messages.properties
@@ -1,6 +1,4 @@
 label.regform.title=User Registration
-label.user.firstName=First Name
-label.user.lastName=Last Name
 label.user.email= Dell Email
 label.user.password=Choose a Password
 label.user.confirmPass=Confirm Password
@@ -10,6 +8,6 @@ user.exists=User already exists with given email
 label.resetPassword=Reset Password
 label.resetPassword.link=Reset
 label.pwdform.submit=Update
-label.forotPwd.link=Forgot Password
+label.forotPwd.link=Forgot Password?
 label.comment=Comment
 label.comment.submit=Add Comment
