+++ ideaforce/src/main/resources/templates/forgotpassword.html
@@ -9,7 +9,7 @@
 </div>
 
 <form th:action="@{/user/resetPassword}" method="post">
-    <div><label> Your Dell Email Id : <input type="text" name="email"/> </label> <input type="submit" value="Reset"/>
+    <div><label> Dell Email : <input type="text" name="email"/> </label> <input type="submit" value="Reset"/>
     </div>
 </form>
