+++ ideaforce/src/main/resources/templates/updatepassword.html
@@ -1,7 +1,7 @@
 <html xmlns:th="http://www.thymeleaf.org"
       xmlns:sec="http://www.w3.org/1999/xhtml" xmlns="http://www.w3.org/1999/xhtml">
 <body>
-<div sec:authorize="hasAuthority('CHANGE_PASSWORD_PRIVILEGE')">
+<div sec:authorize="hasAuthority('CP_PRIVILEGE')">
     <head>
         <span th:include="common/header :: common-header"></span>
         <link rel="stylesheet" type="text/css" th:href="@{/assets/css/index.css}"/>
